CRUD Categories finish

+++ app/Http/Controllers/UserController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\UserRequest;
 use Illuminate\Http\Request;
 use App\Models\User;
 use App\Models\Role;
+use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
@@ -19,6 +20,9 @@ class UserController extends Controller
      */
     public function index()
     {
+       if(Auth::user()->role->name != 'Admin'){
+            return redirect('home')->with('error','No puede acceder a este recurso');
+       }
         // $users = User::all();
         $users = User::paginate(10);
         return view('elements.users.index')->with('users',$users);
@@ -30,6 +34,9 @@ class UserController extends Controller
      */
     public function create()
     {
+        if(Auth::user()->role->name != 'Admin'){
+            return redirect('home')->with('error','No puede acceder a este recurso');
+       }
         $roles = Role::all();
         return view('elements.users.create')->with('roles',$roles);
 
@@ -41,6 +48,9 @@ class UserController extends Controller
      */
     public function store(UserRequest $request)
     {
+        if(Auth::user()->role->name != 'Admin'){
+            return redirect('home')->with('error','No puede acceder a este recurso');
+       }
         $user = new User;
 
         $user->fullname = $request->fullname;
@@ -68,6 +78,9 @@ class UserController extends Controller
      */
     public function show(string $id)
     {
+        if(Auth::user()->role->name != 'Admin'){
+            return redirect('home')->with('error','No puede acceder a este recurso');
+       }
         $user = User::find($id);
         return view('elements.users.show')->with('user',$user);
     }
@@ -78,6 +91,9 @@ class UserController extends Controller
      */
     public function edit(string $id)
     {
+        if(Auth::user()->role->name != 'Admin'){
+            return redirect('home')->with('error','No puede acceder a este recurso');
+       }
         $user = User::find($id);
         $roles = Role::all();
         return view('elements.users.edit')->with('user',$user)->with('roles',$roles);
@@ -87,8 +103,11 @@ class UserController extends Controller
      * Recibe la solicitud de actualización de un elemento y actualiza el registro
      * PUT
      */
-    public function update(Request $request, string $id)
+    public function update(UserRequest $request, string $id)
     {
+        if(Auth::user()->role->name != 'Admin'){
+            return redirect('home')->with('error','No puede acceder a este recurso');
+       }
         $user = User::find($id);
 
         $user->fullname = $request->fullname;
@@ -114,6 +133,9 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        if(Auth::user()->role->name != 'Admin'){
+            return redirect('home')->with('error','No puede acceder a este recurso');
+       }
         $file = public_path().'/'.$user->photo;
         if (!str_contains($user->photo, 'https') && getimagesize($file) && $user->photo != 'images/no-profile.png') {
             unlink($file);
